git rid of final pep8 exceptions
