+++ hvcc/core/hv2ir/HLangNoise.py
@@ -13,28 +13,32 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import os
-import random
 
-from .HeavyLangObject import HeavyLangObject
-from .HeavyParser import HeavyParser
+# moved to HeavyParser.py because of circular dependency
 
 
-class HLangNoise(HeavyLangObject):
-    """ Handles the HeavyLang "noise" object.
-    """
+# import os
+# import random
 
-    def __init__(self, obj_type, args, graph, annotations=None):
-        assert obj_type == "noise"
-        HeavyLangObject.__init__(self, "noise", args, graph,
-                                 num_inlets=1,
-                                 num_outlets=1,
-                                 annotations=annotations)
+# from .HeavyLangObject import HeavyLangObject
+# from .HeavyParser import HeavyParser
 
-    def reduce(self):
-        seed = int(random.uniform(1, 2147483647))  # assign a random 32-bit seed
-        noise_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "./hvlib/noise.hv.json")
-        x = HeavyParser.graph_from_file(noise_path, graph_args={"seed": seed})
-        x.reduce()
-        # TODO(mhroth): deal with control input
-        return ({x}, self.get_connection_move_list(x))
+
+# class HLangNoise(HeavyLangObject):
+#     """ Handles the HeavyLang "noise" object.
+#     """
+
+#     def __init__(self, obj_type, args, graph, annotations=None):
+#         assert obj_type == "noise"
+#         HeavyLangObject.__init__(self, "noise", args, graph,
+#                                  num_inlets=1,
+#                                  num_outlets=1,
+#                                  annotations=annotations)
+
+#     def reduce(self):
+#         seed = int(random.uniform(1, 2147483647))  # assign a random 32-bit seed
+#         noise_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "./hvlib/noise.hv.json")
+#         x = HeavyParser.graph_from_file(noise_path, graph_args={"seed": seed})
+#         x.reduce()
+#         # TODO(mhroth): deal with control input
+#         return ({x}, self.get_connection_move_list(x))
