+++ tox.ini
@@ -34,6 +34,5 @@ cov-report = html,term
 omit = .tox/*,venv/*,tests/*,examples/*,setup.py
 
 [flake8]
-ignore = E402,W503
 max-line-length = 120
 exclude = .tox,venv,build,examples
